Fix: weight of forged object could sometimes be incorrect.
This one was reported by mightyquinn - if forging an object whose
material was not the same as its base material, its weight would still
be that of its base material. For example - a normal dwarvish mattock
weighs 120 aum. If you forge a pick-axe and a mithril dwarvish short
sword together, you'll create a mithril dwarvish mattock, whose weight
should be 45 aum, but it was still coming out as weighing 120 aum. This
commit fixes this.