Fix: forging and object properties, stacks.
During testing, I wanted to make sure the object properties were being
carried over from either of the recipe objects, that bit of code had
made it in. Realistically, you shouldn't know the object property status
of the newly forged object unless you already knew one of the object
properties of either of the recipe objects. So that's fixed here.

Another thing I noticed - forging a dart and a dagger together was
supposed to only make one shuriken. A stack of shuriken was being made
instead. I'm not opposed to a stack of huriken being produced, so long
as it's not over the top. My test run, I used 100 darts and 100 daggers
to create almost 900 shuriken. That's way too many. I've limited the
stack to 3, with the off chance you can produce a stack of 4.

Last bit - I lowered the chances of causing a forge to cooldown and stop
working when forging objects with magic (object property).