Fix: Vecna wasn't always dropping its artifact.
This wasn't really noticed before when Vecna only dropped the Eye of
Vecna, because there was only a 50% chance of it happening. But
now that Vecna is guaranteed to drop either the Eye or the Hand,
it was noticed when that didn't happen just recently. There's a
define in mon.c called LEVEL_SPECIFIC_NOCORPSE, and if a level or
region is flagged as a graveyard, undead only leave a corpse 33% of the
time. Liches don't generate corpses regardless, but the routine that
generates which artifact Vecna drops is in corpse_chance(), and Vecna
falls under is_undead(). This commit fixes the no artifact drop issue.