Tortle enshelling improvements.
Noticed (myself and others) that when a tortle goes to hide in its
shell, it's not immediate. It can take 1-2 turns before a tortle is
fully in its shell and is fully protected. Meaning that during the
process of going into its shell, its still vulnerable. This is not the
intent - the whole point of hiding in ones shell is purely a defensive
mechanism, and is meant to help save our hero in a crisis scenario. If
the act of hiding isn't immediate, then what's the point? So that's
addressed in this commit, by calling find_ac() in toggleshell(), so that
the tortle's AC boost is immediate.

Another bit I decided to add, based off of player feedback and my own
notes (I initially thought this would be too OP) - even when hiding in
your shell, low level monsters can still periodically land a hit, and a
low level player just can't regenerate hit points fast enough to
compensate, almost becomes a death trap of sorts. So, allow an enshelled
tortle a limited type of hit point regeneration, if not already having
regeneration boosted. It's not as good as wearing a ring of the same,
but it's good enough to keep you alive. May tweak this as necessary.