More Goblin Town tweaks, defeating boss monster feedback.
This is partially based off of player feedback I saw earlier today in
the #evilhack IRC channel. First bit - the prisoners in Goblin Town are
not strictly elves, but a mix of elves, dwarves, and gnomes. If you chat
them, they have specific things to say depending on whether the Goblin
King lives or not. All of them are generated peaceful, and will not have
any possessions.

Second bit - killing the Goblin King, Cerberus, and Vecna: all three of
these events unlocks a specific aspect of the game. Thought it would be
neat to set up a com_pager() event for the death of each of these, which
drops a not-so-subtle hint for what is possible that previously wasn't.
Adds more depth to the game, makes the player feel like they're more a
part of the narrative, rather than just a hack n' slash killing machine.