Cursed wand of make invisible makes target visible.
Zapping a cursed wand of make invisible at yourself or a monster will
cause you/them to become visible if already invisible intrinsically. If
the target is already visible, nothing happens. If changing yourself
from invisible to visble via this method, there's a 50% chance to
aggravate monsters on that level.