Change critera for which body armor is allowed to be worn whilst having wings.
It was recently pointed out that with the addition of dtsund
dragon-scaled armor, a player could wear cloth/leather/dragonhide body
armor infused with scales and still use their wings, but couldn't use
the same if the body armor wasn't infused with scales. That's not right.
Now, any body armor that isn't rigid/hard (made of cloth/leather or
dragonhide) will allow holes for wings to pop through.