Adjust blocked access to Mine Town/Mines' End while Goblin King is alive.
Before this commit, the stairways into Mine Town were blocked, both from
above and below it. Levelporting to circumvent the magical blocks is
allowed, but if the player hit a levelport trap and was placed between
Mine Town and Mines' End, they were effectively stuck without a means of
levelporting. What's the point of using levelporting to get around the
blocked stairs if it does you no good? The change is this - the block on
the stairs leading up into Mine Town from below has been removed, and
replaced with a block on the stairs *in* Mine Town going down below it.
This way, if a player gets levelported below Mine Town (either on
accident or on purpose to get around the magical block) and they haven't
killed the Goblin King yet, they can make their way back up and into
Mine Town and up out of it, they just can't go back down again
physically until the Goblin King is defeated. No more getting stuck in
limbo.