Forge recipe improvements, slight recipe changes.
Change the recipe arrays into a single struct, add the ability to
require more than one of the same object for certain recipes (these need
to be able to stack currently). Altered a few of the recipres to require
more than just one arrow or dart to forge the desired object; changed
two recipes - to forge a javelin now requires crossbow bolts (2) and a
spear, to forge a lance now requires a javelin and a glaive. The
'Blacksmith's Cookbook' has been updated accordingly.

This sets the stage for some other neat recipe ideas I have. Thanks
amateurhour and qt for the feedback, and qt for some of the code to get
me pointed in the right direction.

Co-authored-by: entrez <me@entrez.cc>