Fix: saving versus ball & chain.
With the inclusion of commit b616c1c, saving while punished (ball &
chain) could cause a double free, or trying to free an object that was
already freed from memory previously. I've seen evidence of this via
the paniclog as well as odd coredumps that didn't produce any
meaningful backtraces, but when reviewing the ttyrec of that game,
always showed the player saving/exiting. In NetHack 3.7, this issue is
already addressed - I've incorporated commits f8fcab3 and 4d6a140 which
appears to resolve that.