Eating the Hand of Vecna is bad for you.
Just like eating the Eye of Vecna, eating the Hand is bad for you, and
can kill you if you're not strong enough. The Hand of Vecna isn't
considered an actual food object like the Eye is, but it could still be
possible if poly'd into something like a gelatinous cube.

You already couldn't eat the Hand once it's worn, but I altered the
feedback should someone attempt it.