Fix: certain forge recipes could cause splitobj panic.
When forging objects that called for a small stack of something as part
of the recipe, if that recipe wasn't followed in the eact order as it's
expressed in the array, it would cause a 'splitobj' panic.

Example - forging 2 arrows and a dagger will create a spear. If the
arrows were the base object, then the dagger, no problem. Selecting the
dagger first, then the arrows, would have removed only one of the
arrows, and would leave negative one daggers in the players inventory.
This would trigger the panic, and then eventually a crash.