Make crystal plate mail more rare.
Now that crystal plate mail can be forged, a couple players pointed out
that there wasn't really a need to do so, because it was fairly easy to
obtain off of player monsters or Yendorian army captains. After having
another look at how often that could occur, it looks as if crystal plate
was indeed too easy to get. I also noticed that the probability numbers
for armor in general were skewed - the total probability doesn't have to
total up to 1000, but it was less than that, enough so that armor spawn
rates would be different vs vanilla NetHack. So while adjusting crystal
plate mail spawn chances, I made sure that the total probabily value for
all armor totalled to 1000.