Fix: enchant armor selection and dragon scales; wishing for dragon scale mail.
A couple fixes that revolves around dragon scales:

1) Before dtsund-DSM was implemented, players had the option to select
which piece of armor to enchant if multiple pieces were worn, which is a
nice 'quality of life' feature. The dtsund implementation had it so that
if scales were worn, it bypassed any selection and focused the
enchantment on it alone. This caused a number of issues, especially if
the player already had dragon-scaled armor and were wearing another set
of scales as a cloak. This commit effectively gets rid of the bypass.

2) Another quirk of dtsund-DSM - wishing up '<foo> dragon scale mail'
would produce nothing. Twice now, I've watched players not aware or
not familiar with dtsund-DSM try to wish up dragon scale mail and burn a
number of wishes trying to get it, only to wind up with nothing. They
thought they had misspelled the wish or something. Frustrating as hell
to say the least. Now, if the player wishes for '<foo> dragon scale
mail', they'll receive 'a set of <foo> dragon scales' instead. They may
still be a bit confused if they're unaware of the dtsund changes, but at
least they'll pretty much get what they wished for. This will catch the
muscle memory wishes as well for those who know about dtsund-DSM and
just forgot/are on autopilot.