Switch heavy war hammer to regular war hammer (Goblin Town).
Artifacts like Mjollnir and Ogresmasher use heavy war hammer as their
base object. Sometimes when the heavy war hammer would spawn in one of
the chests found in Goblin Town, it would be one of these artifacts -
not the intent. Switch it to just a regular war hammer, which is more
than sufficient to forge with.