Allow barding to be forged.
Community suggestion to allow barding to be forged, which makes so much
sense. Kinda surprised I didn't already think of it. All three types of
barding can be forged. If one of the recipe objects has an object
property, don't allow it to carry over to the newly forged barding (it's
not allowed already, check needed to be put in place to prevent some odd
possibilities). The Blacksmith's Cookbook has been updated accordingly,
and I adjusted some entries so some pages would display all on one page
on your typical 80x24 terminal display.