Fix: all unlocking tools being treated as magic when unlocking a trapped container.
The issue - player goes to unlock a large box or a chest, said container
is trapped. The player unlocks the container, and it automatically goes
to the 'You find a trap!  Do you want to disarm it?' routine, answering
yes successfully disarms the trap every time. This is only supposed to
happen when using the Master Key of Thievery, but instead was happening
with any unlocking tool. Turns out this is an old bug from at least
NetHack version 3.6.x, maybe even older. The solution lies in how
is_magic_key() is expressed (see NetHack git commit 86120d4).

While fiddling around with this, I did a bit of code cleanup in lock.c,
and discovered/fixed a minor bug regarding the special 'trapped' doors
at the end of Sokoban. You can't untrap the 'trapped' doors where the
prizes are held, but if you had tried to untrap a trapped container in
Sokoban, you would have received the same feedback and failed.