Fix: reading scroll of enchant armor while confused and wearing dragon scales.
If reading a scroll of enchant armor while confused while wearing a set
of dragon scales, and then selecting a different piece of worn armor to
fooproof, the selected armor would enchant instead of fixing it and
removing any erosion/corrosion. This also gets rid of a boolean that is
no longer used.