Create is_open_air() function.
xNetHack borrowed the idea of having areas of open air where gravity was
in effect and the player/monsters/objects could fall. The implementation
there is different than what was made for EvilHack. One nice function
that was created in xNetHack, was is_open_air(), which makes the code
look nicer and makes it easier to read, and for EvilHack, gets rid of
the Vlad's tower specific instances. This sets the stage to make it
trivial to edit/create other maps with open air and have the same
effects apply.