Fix: prevent player from hiding their quest artifact in a container to avoid interaction with quest leader.
It was discovered that you could sneak by your quest leader with the
quest artifact by placing it inside a container. This would be done if
the player thought that the quest leader would demand that they hand
over the quest artifact. This would just put off the inevitable, as the
quest would still not be marked as complete, and the Bell of Opening
will not work. But, sneaking away with the quest artifact was never the
intent.

This commit is a quick fix. There may be a better way to do this, such
as the quest leader sensing you have the quest artifact in your
possession inside your bag, and forcing you to remove it so they can
examine it. But for now, this'll work.