Fix: pushing boulders, ball & chain handling in Vlad's cavern.
This commit incorporates various recent commits to xNetHack
Significantly modified), which also has air terrain in Vlad's cavern
that the player can fall from if not flying or levitating. The bugs
that are addressed: the player could push a boulder out into air
terrain without it falling like other objects. Ball & chain would
follow the player on air terrain (it still will whilst levitating, as
it's supposed to). This also addresses a crash that could occur if
punished and unpunishing yourself while on air terrain (ball & chain
would not be properly freed).