Fix: shallow water not drying up from objects being dropped into them.
Quite a ways back, a condition was made so that you couldn't endlessly
dip objects into a shallow pool/puddle. Eventually it would dry up. It
was recently discovered that this could be bypassed by dropping objects
into a shallow pool - same effect as dipping, but without drying up the
puddle. This commit addresses this.