Fix: metalvores eating a metal container would also consume all contents.
Was recently noticed that if a metalvore, such as a Xorn, were to eat a
container made of metal, all of the contents of the container would go
away, whether they were made of metal or not. This commit fixes this by
making all of the contents of said container spill out as its eaten,
much like how a metal container will spill its contents if it rusts
away. The metalvore that ate the container could then still go on to eat
any metallic objects that had spilled out of the eaten container.