Monsters that gain resistance to a type of damage also lose any vulnerability to the same damage type.
It was noticed that if a monster (typically a pet) gained a resistance
to something, say fire, and were also flagged as being vulnerable to
fire, the vulnerable flag would persist even though that monster was
immune to its effects. This commit addresses this, plus corrects a bug
in pager.c in the process. Thanks qt.

Co-authored-by: entrez <me@entrez.cc>