Fix: foo, poisoned by a cursed amulet of life saving.
Wearing a cursed amulet of life saving will not work, and the player
will die via whatever caused them to die in the first place. The
inclusion of KILLED_BY here was not entirely correct. Yes, the cursed
amulet of life saving didn't help you, but that's not why you died.

This commit incorporates NetHack 3.7 commit 47cef18, which provides a
proper 'while' statement if an amulet of life saving doesn't work. Then
we call savelife() in the routine that causes a cursed amulet to fail.
End product is proper death reason output.