Fix: memory corruption issues caught with address sanitizer.
First one - zapping certain types of wands in the corner of a map, if
pointed off the map would cause a buffer overflow. amateurhour caught
it, qt provided the fix.

Second - an underflow in free_window_info that was recently fixed in
NetHack 3.7 (commit f9d7bc6). Saw it while working on the first issue.

Co-authored-by: gebulmer <gebulmer@googlemail.com>
Co-authored-by: entrez <me@entrez.cc>