Make berserkers more willing to fight hand-to-hand
Berserk monsters won't be as interested in hanging back and fighting at
range, and instead will be more likely to close the distance to the hero
and attack hand-to-hand.  I think this is sufficient to make this
happen, though it will need some testing to ensure monsters properly
switch to hand-to-hand weapons as they get close -- it looks like that
should happen automatically in mattacku, though.