Fix: sections of wall being visible when they shouldn't yet.
This has been a long-standing bug for as long as I can remember, and qt
appears to have figured it out. What was happening: the player would all
of the sudden see a section of wall in an area that they hadn't explored
yet. It was discovered that this was only occurring if that section of
wall had any type of tree up against it. This bug might also be present
in vanilla NetHack, but it's unclear because trees don't randomly spawn
in the dungeon like they can in EvilHack. IS_TREES() covers both normal
trees as well as dead ones.