Mimics and dead trees.
Mimics could mimic being a dead tree, but it wouldn't block what was
behind it (dead giveaway), so fix that. Also set a condition that if a
mimic spawns inside a garden room, it will mimic either a tree or a dead
tree.

Discovered that dead trees weren't tracked under seen objects per map,
so that's been corrected (this adds extra info to the 'mapseen' struct).

Last bit - some unrelated formatting cleanup in mhitu.c