Allow normal sight whilst underwater given the right conditions.
If our hero is underwater, and they are poly'd into a monster that is
amphibious, or if they have intrinsic magical breathing, they can see
farther than the typical one tile. They'll be able to see as far as what
would normally be allowed if they were on dry land. The tortle player
race is amphibious naturally, therefore this ability is default for
players playing as a tortle. Monsters that are normally hidden in water
are not under these conditions.

Currently there is one item that will also allow this type of sight -
wearing the Eyes of the Overworld. I may create something else that
allows for this, perhaps something like a diving helmet. We'll see.

This work is done in preparation for an upcoming optional side quest
that is centered around an underwater adventure. There may be more to
tweak with this, but this lays down the groundwork for now. Thanks qt
for making this possible.

Co-authored-by: entrez <me@entrez.cc>