To-hit calculation on the bottom line.
Pulled from SpliceHack git commit cdd9a9c (modified).

Shows the base to-hit calculation on the bottom line, next to the AC
value. The higher this value, the better the chances are for the player
to land a hit. This value can be negative (e.g. a monk wearing body
armor). Racial modifiers, or adjustments that come into play based on
the target monster aren't calculated in this to-hit display, since
there's no way for it to know what you intend to attack. However, much
is taken into account here, so you'll have a pretty good idea of what
your chances are to-hit based on the weapon used and its enchantment,
your stats, your skill with the weapon wielded, exp level, luck, and so
on.