Magic negation (MC) value on the bottom line.
Got the idea from SlashTHEM, but this is my own implementation. This
allows the player to see what their MC rating is on the bottom line.
Previously, the only way you could know for sure what your MC was, was
by using enlightenment. Thought this would be fun for the player to see
at a glance, and not have to guess which combo of
armor/rings/amulets/artifacts/divine protection was providing MC.