Allow looting of containers that are in water.
Allows you to loot containers that are on the ground while underwwater,
lava is still off-limits. This does not have any handling for water
rushing into a container if it's opened - vanilla NetHack doesn't have
any handling for this either, most likely because the devteam never
really expected it to be a common occurrence. So this may be something
that's added later on.