New object: goggles.
Goggles are worn in the same slot as lenses or a blindfold, and allow the
wearer to see at a distance whilst underwater. They can protect your
eyes under certain circumstance just as regular lenses can. Goggles are
made of plastic and are black in color. They are non-magical, but are
more rare than lenses. Can be wished for.

Also corrected an oversight from last commit (looting containers in
water).