Fix: tipping into containers.
This commit incorporates NetHack 3.7 commits 7ae4efb, d0ee89a, c2a9343,
5b0f4a2, and aa8473c (modified). Tipping objects like wands of
cancellation, bags of holding or tricks into another bag of holding will
cause the receiving BoH to blow up. Made a tweak to Magicbane - if its
enchantment is 0 or less, it will not potentially cause a BoH to blow up
if it's put inside of it. There's handling here for tipping a horn of
plenty also.

Tipping into containers on the floor isn't implemented here yet. I
imagine the NetHack devteam will get to that eventually, so as they make
progress towards that, I'll incorporate their changes.