Fix: to-hit display weirdness under certain conditions.
If twoweaponing and your offhand weapon was too heavy, or if using an
edged weapon as a priest, the feedback given while doing these things
would display twice, with the second display showing in the middle of
the map and screwing up the display. This was because the feedback for
these things resided in weapon_hit_bonus(), and with the inclusion of
the to-hit display, it was being called twice. Moved the feedback into
attack() in uhitm.c