Fix: to-hit display revealing weapon enchantment, other tweaks.
The to-hit display was revealing a wielded weapons enchantment. Fix
this, so the to-hit display will only take a weapons enchantment into
account if the enchantment is known, otherwise it will be treated as no
enchantment. Need to figure out a way to handle this for un-id'ed rings
of increase accuracy as well, haven't cracked that nut yet.

This commit also tweaks the to-hit value formula so it matches more
closely to the actual to-hit, and also some more (and hopefully final)
fixes for twoweaponing feedback.