Weapon familiarity.
Added this based on player feedback, pulled it from SpliceHack. Couldn't
find an actual commit for it, but it was easy enough to find the
necessary code.

When attacking with your primary weapon, you have a chance to ID its
enchantment after a certain number of kills. The more powerful the
enchantment of the weapon, the less chance there is of ID'ing it. Keep
swinging it enough times and rack up the kills, you'll eventually get a
successful enchantment ID.

There's some unrealized potential here. First thing that comes to mind -
receiving Stormbringer as a crowning gift, the player is chosen to
'steal souls' for the glory of their deity. Could definitely do some
neat things with that now that wep_kills is a thing that can be tracked.