Goblin Town expansion: Gollum's Cave.
In J.R.R Tolkien's book 'The Hobbit', Bilbo had found himself below
Goblin Town, and stumbled upon Gollum's cave. Thought it would be neat
to include this as an addition.

The only (non-magical) means of making your way down into Gollum's cave
is via a trap door/hole that is randomly placed in Goblin Town. Falling
through lands you on a narrow rocky ledge that is the shore line for a
hidden lake underneath the town. This is where Gollum lives - he's lost
his 'precious' and is searching for it. Gollum's 'precious' is a cursed
ring of invisibility named 'The One Ring' that is randomly placed
somewhere on this level, and it will behoove the player to find it
before Gollum does. There's a small main island in the middle of the
lake, along with a couple other smaller islands dotted around the level.
There's enough shallow water about, the player could walk to the main
island if they so chose. There is a stairway up in the southwest part of
the map that will bring the player back up near the exit portal in town.

There are other inhabitants here other than just Gollum, of which he
will attack and eat. Gollum is not very strong, but can steal single
objects and will try to run off with them.

Gollum's Cave is really more about adding a bit of flavor and content to
Goblin Town than anything else. There's nothing here that the player
needs to accomplish to open up other parts of or win the game. Just
thought it'd be neat to include a bit more Tolkien lore, and possibly be
a fun detour.