Sanctum/ascension run revamp: part 1 (sanctum).
This is the first part of the sanctum/ascension run revamp for EvilHack.
The plan is this: instead of the player having to trek all the way back
out through Gehennom and the dungeon to escape and enter the elemental
Planes, they will instead make their way from the bottom of Gehennom
(the Sanctum) up through Purgatory, and then be deposited in the dungeon
on the same level as their quest home entrance (this is in case the
player needs to chat with their quest leader again). They then traverse
the few remaining dungeon levels to the Planes. Purgatory will only be
2-3 levels at most, but will be just as difficult as the Sanctum, if not
more so. It won't be an easy trek back out to the Planes, but it will be
a much shorter and less tedious one.

Part one focuses on the Sanctum, and the west side of the map has been
redesigned to accommodate the changes. Once the player acquires the
Anulet of Yendor (for Infidels, imbue the Idol of Moloch), a gate is
formed in the west wall of the inner sanctum, leading to a previously
unreachable part of the map. At the same time that this gate magically
appears, the stairs leading back up out of the sanctum are removed, and
all other forms of leaving the sanctum (branchporting, levelporting, or
quaffing a cursed potion of gain level) are suppressed. The only way to
leave the sanctum and continue is through the newly formed gates, and to
face the monsters within.

In this new section resides Lucifer, the most powerful demon prince of
them all. He guards the magic portal that leads to Purgatory. His
entourage are four named balrogs waiting outside his inner chamber.
Lucifer will not engage until the player attacks or moves next to him.
His attacks, while not as severe as Demogorgon's illness attack or
Graz'zt's stealing ability, are stong enough to be taken seriously.
Lucifer also spawns with 666 hit points, so he won't be an easy or quick
kill, even with silver weapons.

There's room for improvement here, such as the com_pager() dialogue, or
what gear Lucifer could spawn with, and so on. I added the branch port
location to Purgatory in the .des file, but as of this commit, it isn't
active yet since the Purgatory branch hasn't been created. Once
Purgatory is created - if the player needs to go back there for any
reason once they've exited and are back in the regular dungeon, they
will have the ability to travel back down to the bottom of Gehennom and
enter the Sanctum via the stairs formed during the invocation, and then
through the portal to Purgatory again.

In this commit, there's also a bugfix concerning u.uachieve.amulet and
when an Infidel actually imbues the Idol of Moloch.