Sanctum/ascension run revamp: part 2 (Track Lucifer alive/dead status, foundation for Purgatory levels).
Part two of the sanctum/ascension run revamp. This commit focuses on a
few things, mainly tracking Lucifer's alive/dead status, and adding a
foundation for the Purgatory levels. For now it'll just be two level in
Purgatory, but I may add a third, depends what else I come up with. So,
things in this commit:

* Purgatory levels created (placeholders only, this is NOT final design)
* Track whether Lucifer is dead or alive, defeating him is flagged as
an event and an achievement
* Track entrance into Purgatory, flagged as an event and an achievement
* The portal to Purgatory will not function until Lucifer is killed
* The portal leaving Purgatory puts the player in the first level of
sokoban (explanation below)
* The portal from Purgatory to sokoban only works one-way, meaning you
can't enter Purgatory from sokoban

I initally wanted the exit portal from Purgatory to drop the player off
at their home level for their quest, but the way the level names work
for the quest, I don't think there's a way to do that, so sokoban was
plan B. In each sokoban starting level, I made a little closet where the
exit portal resides - it won't interfere with solving that sokoban
level.

The two purgatory levels are just placeholders, I wanted to make sure
the functions I was setting up would work the way I wanted them to. The
final level designs will be neat.

There are some minor fixes in here as well from when I added Gollum's
Cave to the goblin king branch, I neglected to include that level when
building the game on other OS'es other than *nix.