Lucifer tweaks.
It was pointed out that Lucifer in his initial configuration makes it
virtually impossible for a player to attempt the pacifist conduct. No
monster would ever be of a high enough level to want to attack him, and
even if they did (conflict) they'd never land a hit. Since defeating
Lucifer is required to progress and subsequently win, some changes were
in order.

* Lower Lucifer's default base level and base AC
* Grudge condition set up - if you have a pet, they will attack him
* In the off-chance you encounter Lucifer before obtaining the Amulet of
Yendor, prevent him from warping to the upstairs (as far as I can tell,
this scenario can never happen, but we'll cover our ass here)
* Change Lucifer's size from large to huge
* Give Lucifer a morning star as a weapon. Get it... ?

Lucifer is still the strongest and toughest of all the demon bosses.
Between his max hit points and damage output, you'll still need to come
prepared for a serious fight. Pacifists have a fighting chance now.