Fix: shambling horror pokedex lookup.
The shambling horror is different every single game, you never know
exactly what you're going to get until you encounter one. Or at least
that's how it used to be, until the addition of the pokedex. Some have
discovered that you could simply issue '/? shambling horror' at the
beginning of every game and know their stats. This was never the intent
- there's supposed to be that uncertainty and fear factor associated
with encountering one. This commit fixes this, so when looking one up in
the pokedex now, you'll simply receive its encyclopedia entry.