Place Fort Ludios portal in the first eligible vault generated.
Initially from UnNetHack, but other variants have done this as well. The
portal to Fort Ludios will generate within the first gold vault past
dungeon level 10, as long as the default criteria are met. There is
still a small chance that the portal to Fort Ludios won't spawn, but
that could only happen now if a gold vault is never generated between
dungeon levels 10 and the Medusa level. The odds of that happening are
very low, roughly 5% of all games played.

In UnNetHack, there is a commit that allows the portal to Fort Ludios to
appear on Medusa's level if a gold vault never appeared after dungeon
level 10. I might add that, but for now this will suffice.