Wielding Vorpal Blade protects against decapitation attacks.
For a long time I thought this was already a thing in vanilla NetHack.
Kinda surprised it isn't. In this commit, if the player or a monster
wields Vorpal Blade, they are protected from being decapitated (any
monster that has AD_BHED damage type). The feedback is the same as if
they didn't have a head (wildly miss).

This commit also fixes a couple minor bugs I noticed with AD_BHED
feedback, and some tortle Hidinshell bits that were causing a blocked
attack to stop there and not go on to the next attack in the chain.