Fix: logic with monsters recharging wands.
This addresses how an intelligent monster would determine which
wand/horn it should recharge if it had the means to do so, and in which
order. This also fixes a bug discovered that would cause a monster to
change a wand/horn a second time while it still had charges left to use.

Co-authored-by: entrez <me@entrez.cc>