Fix: monster recharging a wand of wishing.
For the longest time, I had made the assumption that a monster could
recharge a wand of wishing as it would for offensive wands. They can in
GruntHack, but the way it's done is completely different. This commit
allows monsters to actually recharge a wand of wishing, and sets the
foundation to be able to recharge other types of miscellaneous objects
in their inventory.