Fix: various issues with monster wishing.
Some of this comes from the yet to be released variant HackEM, which
uses EvilHack as a base. While merging those fixes, I noticed another
issue with wished for wands not having any charges. I also switched
'xname' to 'ansimpleoname' when displaying what object a monster wished
for - this adds a bit of uncertainty to what exactly was wished for
unless the player has already ID'ed that object. It also handles a minor
feedback bug when describing multiple objects.