Purgatory map creation.
This commit is just making the map layout for the Purgatory levels.
Monsters, loot, etc still need to be added. And these designs may not be
final, but they're close.

The final Purgatory level has open air, just like Vlad's cavern. A tweak
was made in hack.c so auto travel won't send the player over the edge
while in Purgatory.