Finish first Purgatory level.
In this commit, the first level of Purgatory has been finished. This is
where our hero first enters Purgatory, and even though their
surroundings look tranquil, they will soon discover this is not the
case. There are some other changes in this commit to compliment
Purgatory. The breakdown:

* The first Purgatory level, the player will encounter several player
monsters who are trying to escape. They are surrounded by the souls of
past player monsters (spectres) who despise these heroes and will attack
them
* Other monsters reside in Purgatory as well, with a mix of random
spawns that will be different every encounter. The type of monsters that
can spawn here has been limited to fit the theme
* The Wizard of Yendor is guaranteed to appear upon first entrance to
Purgatory. Throughout the course of the ascension run, there's now a
greater chance for Rodney to appear, since the entirety of the ascension
run has been greatly shortened
* Since player monsters will do battle with the spectres here, their
weapons need to be able to land a hit, and during testing I noticed that
most of them could not. So I've implemented a change to shade_glare()
that allows any weapon with a +2 or greater enchantment to be able to
hit non-corporeal creatures. This is a change several of us have
discussed in the past anyways, and this was as good a time as any to
make it a reality
* While in Purgatory, hostile player monsters will shout various
hostilities at the player in regards to the Amulet/Idol

Is there room for improvement here? Probably. Good start though. Next up
is the final Purgatory level.